// resources/js/api/generated/models/volumeInfo.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { VolumeInfoLabels } from './volumeInfoLabels';

export interface VolumeInfo {
  created: string;
  driver: string;
  labels: VolumeInfoLabels;
  mountpoint: string;
  name: string;
  size: number;
  unused: boolean;
}